bpf: actually pin the HID_BPF link, not the program
The fact that the program is attached to the device is now represented
by a link, which is returned by hid_bpf_attach_prog().

Pinning the program itself doesn't maintain that link alive, and the
program gets unattached as soon as the rust executable finishes.

This allows to check if
- the syscall was correctly called
- the hid_bpf_attach_prog was successful
- the hid_bpf link has been properly pinned

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@gmail.com>