main: fall back to filenames if our udev device fails
If we get ENODEV for a libudev device during remove, assume the device
has been removed and analyse the provided syspath directly. The last
element is our sysname so let's use that to rm the objects.

This allows us to keep the lookup logic in one place only since we can
now call remove from the udev rules.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>