main: change the example syspath to the device
For adding, the hidraw path works fine. For removing the hidraw path
would work fine too but since remove is likely called when the device is
being removed, any symlinks from hidraw to the old path can no longer be
resolved. So let's change the example to the proper hid path.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>