bpf: add metadata to the generated BTF
Following a hack from libXDP, we can tweak BTF to include metadata of
the programs we load.

This is by defining a global variable that is never instantiated, but
which type encodes the parameters through a set of arrays.

Then, when we load the bpf object, we can parse the btf, and extract the
modaliases.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@gmail.com>