bpf: rename debug.bpf.c into trace_hid_events.bpf.c
Now that hid-bpf programs can be attached to any device, we can
make this program more generic: it's not USB related and can be applied
to any HID device.

Also, there is no point in preventing it to load from the probe,
so remove that too.

Remove the .hid_bpf_config section as libbpf has trouble loading an
empty section.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@gmail.com>