XP-Pen Artist 24: do not be too aggressive with out-of-prox filtering
Turns out that the previous code was also filtering regular out-of-prox
events, because there was no reason for the tilt to be set to 0 before
the pen leaves the sensor proximity.

Now only strip out out-of-prox events when the pen is in contact.
We know that the device is reliable regarding in-range, so if a contact
is occurring and we get an immediate out-of-prox, this means that the
secondary barrel switch has been pressed.

Let's hope the device doesn't emulate a full out-of-prox event...

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@gmail.com>