XP-Pen Artist 24: final touches on the filter
The last test (TIP_SWITCH | BARREL_SWITCH | ERASER | IN_RANGE) == 0 was
never true, because we already had a "if ((IN_RANGE) == 0) return" at the
beginning.

If we get an out-of-range and the tip was previously there, filter this
out.

Now we are making the stylus completely compatible without spurious events!

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@gmail.com>