XP-Pen Artist 24: take into account contact + eraser information
The firmware emulates an eraser mode while pressing the second button.
We can detect them when both the tip switch and the eraser are changing
at the same time. This is very highly impossible to achieve by a human,
so this is very likely a firmware glitch.

But given the intrinsic nature of tablets, even if it were a correct event,
we would get one other after with the correct data, so dropping a 3ms
event won't change the face of the world (especially if the user is
interacting with their buttons, so not in a precise environment).

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@gmail.com>