Add --udevdir handling for the hwdb files to install/release scripts
Guess the right udev dir based on the prefix. This is /usr/lib/udev for
a prefix of /usr and /etc/udev for anything else.

That's close enough to what we want to do - system installs all in /usr/*
and custom installs with /usr/local/bin and hwdb files in /etc.

Note that the bpf objects are still always in /lib/firmware

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>