Add udev-hid-bpf install
This installs the bpf.o and a corresponding udev rule. Unlike our
default setup this udev rule doesn't use the hwdb, it uses the KERNEL
property to match on. Usually udev rules are harder to maintain but in
this case the goal is to install a single udev rule for a single file,
so we're all good.

The lib/firmware path honors the given --prefix (if any, otherwise it
defaults to the compiled one). The udev rule is always placed in /etc/
because there are very few use-cases where that isn't the right place to
put it.

Closes #10

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>