udev rule: fix catchall on empty values too
Seen when running `udevadmn test` on any HID device without the `HID_BPF_*`
property set: `udev-hid-bpf add` is called on them.

Turns out that the `ENV{HID_BPF_*}` is matching on the property named
`HID_BPF_*` not any property starting with `HID_BPF_`.

So add a small hint to tell that udv-hid-bpf needs to be called.
We prefix the `.HID_BPF` with a dot because:

> Property names with a leading "." are neither stored in the database
> nor exported to events or external tools (run by, for example, the
> PROGRAM match key).

So the property will not be seen through `udevadmn info` but udev-hid-bpf
will be called.

Fixes #13

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@gmail.com>