bindgen: remove snake_case warning
we conflict between C-style snake_case and rust-style CamelCase for struct
definitions. I used once a `rustbindgen replaces` to sort this out
but the associated test function generated by bindgen was complaining
about not being snake case.

So instead of fighting the compiler, let bindgen translate the snake case
struct and add a type alias at the beginning of the generated file.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@gmail.com>