build.rs: move building bpf files into helper functions
The previous approach was to build attach.skel.c, then
the bpf.o files, then hid_bpf_bindings.rs. This is the natural flow but
but we can build them out-of-order with the two wrappers first (attach
and bindings) and then the bpf.o files. This makes the code a bit easier
to group together.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>