Only consider bpf.o files in $prefix/lib/firmware/hid/bpf
Instead of hardcoding /usr and /usr/local let's work with the
meson-provided bpfdir (from the prefix) and use that as our only default
lookup path.

This means installing in different prefixes will work independently but
that's kinda the point. If there's a use-case for multiple lookup
directories we can sort this out then.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>