tools: use T/S/U as number prefix for the properties
Previously we used a simple numeric count for the properties. This may
mean that where a stable hwdb is installed by the distribution and a
testing hwdb is installed by the user, the HID_BPF_X may overlap with
one of them overwriting the other (depending on hwdb file sort order).

Prefix them so that HID_BPF_T_001 refers to the first one in testing, S
and U for stable and userhacks. This way the conflict is within the same
category only.

This also changes the number to be 001 for better alignment and prints
the .HID_BPF property after the other for a slightly nicer look.

Example entry:

  hid-bpf:hid:b0005g0001v000003F0p0000464A
    HID_BPF_T_000=HP_Elite_Presenter.bpf.o
    .HID_BPF=1

Closes #16

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>