Prefix all bpf.o files with a version number
This allows us to update the BPF programs in the future for newer
kernels but keep the existing ones around for older kernels.
The idea is that once the loader is updated, we will attempt to load
30-foo.bpf.o first, if that fails we'll try 20-foo.bpf.o and if that
succeeds ignore 10-foo.bpf.o (assuming that all three are assigned to a
device).

This patch is the file rename only.

Part of #15

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>