meson: unconditionally recompile any registered bpf file
Turns out that only the initial ninja run compiles everything. When we
do a change in a .bpf.c file not in `-Dbpfs`, meson doesn't detect it
and thus doesn't recompile it.

Ensure the `-Dbpfs` flag (and filter) matters only for the installation.

This allows users who wants to create a userhack to create it, and
recompile it without requesting the full `-Dbpfs=userhacks` to be set,
with the risk of installing unwanted files.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@gmail.com>