81-hid-bpf.rules: always set --bpfdir
The code currently aborts if `--bpfdir` is not provided and if
`DEFAULT_BPF_DIRS` doesn't exist. This is because if this directory
doesn't exist, we can not really load anything from it so we better
bail out.

In the general `meson install` case, this is fine because meson knows
about the actual `$PREFIX` used to install the target.

But in the `make_release.sh` case, `meson` is given a temporary
directory as a prefix and thus has a poisoned `DEFAULT_BPF_DIRS`.

To ensure we always use the installed data, enfore `--bpfdir` in the
udev rule so we don't get any surprises.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@gmail.com>