test: update the script to test udev-hid-bpf add $device
We have test-path-load which tests with a specific bpf.o file.
The previous implementation of test-udev-load used udev's RUN command to
test but let's replace that with a manual `udev-hid-bpf add` instead
(i.e. one without a file).

This makes debugging a lot easier since we get all the --verbose output
right there instead of having to dig it out of the journal.

The udev trigger implementation is now available as test-udev-trigger.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>