bpf: if we fail to pin our maps, remove the bpfs again
Since we can't guarantee correct behavior if the maps don't pin, let's
unroll. Change load_progs() to return a vec of successfully attached
progs and use that to unroll our load if the pin map failed.

The end result is that we remove the same directory as what is the bpffs

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>