add/list: reinstate a default bpf directories list
Currently, each install of `udev-hid-bpf` is hiding any other installation
path for finding their bpf object files.

This is problematic when we need users to test a branch, because this way
we hide any system-wide installation of `udev-hid-bpf` and any bpf object
files shipped with it.

Supply the default list of BPF lookup directories from meson but make
them configurable. Use the most common directories as defaults but users
can override those if they need to.

This way, a manual reinstallation should still be able to reach the
system-wide installation while keeping the new files around.

Closes #20

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>
Signed-off-by: Benjamin Tissoires <benjamin.tissoires@gmail.com>