bpf: add a driver for the Huion Pro Pen tablet (H610)
Requires https://github.com/whot/huion-switcher for switching the tablet
into raw mode first, and blacklist hid-uclogic, so this is more of a PoC
than anything that needs to be merged.

Thus, this is a userhacks file, as it shouldn't be shipped in
distributions.

But we now get the same level of support than hid-uclogic without the
kernel driver

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@gmail.com>