CI: fix make_release for fedora 37
There is a fix not available in fedora 37 to allow compiling bpf.c files
under gcc.

Manually monkey patch it here.

Given that the tree is not clean anymore, the `--dirty` suffix should be
appended to the archive, but given that we know what we are doing, we
don't do that.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@gmail.com>