userhacks: add the QinHeng Foot Pedal as userhack
This is a cheap foot pedal that is sold as rebranded device under e.g.
the iKKEGOL brand. It's a one-trick pony and that trick is "b", unless
you have access to the Windows software to change what the device sends.

Reverse-engineering the configuration protocol doesn't look particulary
hard but it's more effort than spending 15 minutes to write a BPF that
maps the pedal to the keyboard shortcut we want it to.

In userhacks because I'm assuming not everyone want's this thing to send
Ctrl+C but it illustrates how easy it'd be to modify.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>