bpf: inject udev properties in bpf objects
Once the bpf object is loaded, we can fetch the `.bss` (uninitialized
global variables) and `.data` (uninitialized but not const global data)
maps. Those maps contain the global variables in our bpf object.

With a little of BTF introspection, we can detect if the variable name
starts with "UDEV_PROP_", and if so match the rest of the name with the
current udev properties attached to this device.

This makes the bpf object in charge of which udev properties it needs,
and which size each needs to be.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>
Signed-off-by: Benjamin Tissoires <benjamin.tissoires@gmail.com>