bpf: applied a bunch of checkpatch.pl fixes
While preparing the patch series to send those testing files upstream
it came out that checkpatch was really not happy.

This should be only cosmetic.

Note: the hid_report_helpers.h file still complains a lot about
"ERROR: Macros with complex values should be enclosed in parentheses",
but give that the macro is to create fields in an array, I can't find
an easy way to paper over this error.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@gmail.com>