meson.build: add a check for __builtin_memcpy/memcmp
We need that function and older clang doesn't seem to have memcmp, so
let's bail out correctly.

meson needs a compiler object to check if something compiler but
bpf-compiler may be set to something other than $CC so let's do a
double-check for the two being the same and only test if that is indeed
the case.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>