hidudev: move the ignore property check into a helper
And call that helper from cmd_add(). This slightly changes the
functionality: previously HID_IGNORE_DEVICE was only honoured for the
our lookup of matching BPF files but not for loading BPFs provided
on the commandline.

That behavior may be useful for some debugging but realistically it's
better to expect the caller removes that udev property so we have
predictable behavior.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>