Move distinguishing between no path or a path into the CLI handler
Let's make the implementation itself simpler by exposing methods how to
look up a path and loading a set of files. The logic of when to call
which is now moved into cmd_add() which is where we have all the
information about "do we have a path or not".

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>