test: run pytest with xdist if available
If the machine has xdist installed, use it. This can speed up the
python tests significantly though right now we have so few of them that
it won't matter much.

This defaults to -n auto so until the recent changes in the gitlab CI
(which limit the number of visible CPUs per job) it's a bit impolite
since it would fork off the max CPUs. Unfortunately meson cannot access
FDO_CI_CONCURRENT so the polite solution here is to just not install
xdist in our CI runs.

This could also be worked around with a custom pytest runner script that
meson invokes instead.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>