test: make the bpf loading part of each class
Parametrize our test classes with the set of source files we have.
This makes use of our bpf fixture to load the actual object based on a
"source" argument. This allows us to easily add 00020 and similar to the
sources where it applies.

As a side-effect this also adds the current source file to the pytest
arguments which makes it alot easier to debug which file a test failed
on.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>